requirements.txt
flask==2.3.3
flask-login==0.6.2
flask-wtf==1.1.1
email-validator==2.0.0
werkzeug==2.3.7
python-dotenv==1.0.0
wtforms==3.0.1
pillow==10.0.0
bcrypt==3.2.0
itsdangerous==2.1.2
pyjwt==2.8.0
flask-migrate==4.0.4 